Serial mutex and updated pins for new boards

+++ touCAN/Valeo_ext.h
@@ -33,10 +33,10 @@ uint8_t PrepareFHybrid(
   uint8_t* data,
   uint8_t size,
   uint8_t vscem_req_gp_counter,
-  float em_operating_mode_req_ext,
-  float em_voltage_dc_link_req_ext,
-  float em_speed_request_ext,
-  float em_torque_request_ext);
+  double em_operating_mode_req_ext,
+  double em_voltage_dc_link_req_ext,
+  double em_speed_request_ext,
+  double em_torque_request_ext);
 
 uint16_t ArrangeWMHEV(
     uint8_t* dst_p,
@@ -47,13 +47,13 @@ uint8_t PrepareWMHEV(
 	struct x8578_can_db_client_pcm_pmz_w_mhev_t* src_p,
   uint8_t* data,
   uint8_t size,
-  float bisg_calibration_id_req,
-  float em_torque_gradient_pos,
-  float engine_torque_ripple,
-  float em_torque_gradient_neg,
-  float em_cur_dc_link_motor_limit,
+  double bisg_calibration_id_req,
+  double em_torque_gradient_pos,
+  double engine_torque_ripple,
+  double em_torque_gradient_neg,
+  double em_cur_dc_link_motor_limit,
   uint8_t em_motor_limit_grp_counter,
-  float em_vol_dc_link_motor_limit);
+  double em_vol_dc_link_motor_limit);
 
 uint16_t ArrangeTMHEV(
     uint8_t* dst_p,
@@ -64,11 +64,11 @@ uint8_t PrepareTMHEV(
 	struct x8578_can_db_client_pcm_pmz_t_mhev_t* src_p,
   uint8_t* data,
   uint8_t size,
-  float toc_pump_dr_req,
-  float em_coolant_pump_dr_req,
-  float em_cur_dc_link_gen_limit,
+  double toc_pump_dr_req,
+  double em_coolant_pump_dr_req,
+  double em_cur_dc_link_gen_limit,
   uint8_t em_gen_limit_grp_counter,
-  float em_vol_dc_link_gen_limit);
+  double em_vol_dc_link_gen_limit);
 
 uint16_t ArrangeUMHEV(
     uint8_t* dst_p,
@@ -80,26 +80,26 @@ uint8_t PrepareUMHEV(
   uint8_t* data,
   uint8_t size,
   uint8_t vscem_torq_lim_gp_counter,
-  float em_torque_min_limit,
-  float em_torque_max_limit,
-  float drvline_damp_torq_min_lim,
-  float drvline_damp_torq_max_lim);
+  double em_torque_min_limit,
+  double em_torque_max_limit,
+  double drvline_damp_torq_min_lim,
+  double drvline_damp_torq_max_lim);
 
 uint8_t PrepareBCM(
 	struct x8578_can_db_client_bcm_pmz_a_t* src_p,
   uint8_t* data,
   uint8_t size,
-  float car_mode_hs,
-  float car_mode_hs_ub,
-  float power_mode_ub,
-  float power_mode);
+  double car_mode_hs,
+  double car_mode_hs_ub,
+  double power_mode_ub,
+  double power_mode);
 
 uint8_t PrepareGWM(
 	struct x8578_can_db_client_gwm_pmz_h_t* src_p,
   uint8_t* data,
   uint8_t size,
-  float crash_status_rcm,
-  float crash_status_rcm_ub);
+  double crash_status_rcm,
+  double crash_status_rcm_ub);
 
 #endif
