+++ touCAN/dataBroker.cpp
@@ -17,21 +17,21 @@ BrokerData::BrokerData(){
       last_update_time = -1; //data not valid
 }
 
-BrokerData::BrokerData(float initial_value){
+BrokerData::BrokerData(double initial_value){
       this->value = initial_value;
       last_update_time = -1; //data not valid
 }
 
-void BrokerData::setValue(float new_value) {
+void BrokerData::setValue(double new_value) {
       this->value = new_value;
       this->last_update_time = esp_timer_get_time();
 }
 
-float BrokerData::getValue() { 
+double BrokerData::getValue() { 
       return (this->value);
 }
 
-float BrokerData::getValue(int64_t* time_Variable) { 
+double BrokerData::getValue(int64_t* time_Variable) { 
       *time_Variable = this->last_update_time;
       return (this->value);
 }
